legacy_python/requirements.txt
facebook-wda>=1.4.6
tidevice>=0.10.0
aiohttp>=3.8.0
loguru>=0.6.0
paramiko>=2.11.0  # For SSH control of R4S router
Pillow>=9.0.0     # For screenshot handling
mysql-connector-python>=8.0.0
redis>=4.0.0
urllib3<2.0.0
